+++ CHANGELOG.md
@@ -7,6 +7,20 @@ This project follows [Keep a Changelog](https://keepachangelog.com/) and [Semant
 
 ### Added
 
+-
+
+### Changed
+
+-
+
+### Fixed
+
+-
+
+## [2.3.12] - 2025-12-18
+
+### Added
+
 - Added warnings count to scan usage stats.
 
 ### Changed
@@ -19,7 +33,7 @@ This project follows [Keep a Changelog](https://keepachangelog.com/) and [Semant
 
 ### Fixed
 
--
+- Fixed false positive secret detection for certain harmless attribute keys in codebase scanner.
 
 ## [2.3.11] - 2025-12-13